requirements.txt
PyJWT==1.7.1
pip==21.1.2
wheel==0.36.2
pytz==2021.3
requests==2.26.0
yarg==0.1.9
Django==3.2.9
sqlparse==0.4.2
asgiref==3.4.1
certifi==2021.10.8
docopt==0.6.2
pipreqs==0.4.11
urllib3==1.26.7
idna==3.3
setuptools==57.0.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
pansi==2020.7.3
six==1.16.0
redis==4.0.2
Deprecated==1.2.13
py2neo==2021.2.3
interchange==2021.0.4
packaging==21.3
monotonic==1.6
Pygments==2.10.0
pyparsing==3.0.6
wrapt==1.13.3